modifi todo and acout
