+++ todo/app/plugins/SecurityPlugin.php
@@ -74,7 +74,7 @@ $publicResources = array(
     'index'    => array('index'),
     'about'    => array('index','edit'),
 	'register' => array('index'),
-	'todos' => array('index'),
+	'todos' => array('index','edit'),
     'errors'   => array('show404', 'show500'),
     'session'  => array('index', 'register', 'start', 'end'),
     'contact'  => array('index', 'send')
